Restrição de dados Caixa

+++ src/AGENDAHUB/AGENDAHUB/Controllers/CaixaController.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AGENDAHUB.Models;
 using Microsoft.AspNetCore.Mvc;
@@ -17,17 +18,33 @@ namespace AGENDAHUB.Controllers
             _context = context;
         }
 
-        // Função para listar todas as movimentações de caixa
+        private int GetUserId()
+        {
+            var userIdClaim = User.FindFirst(ClaimTypes.NameIdentifier);
+            if (userIdClaim != null && int.TryParse(userIdClaim.Value, out int userId))
+            {
+                return userId;
+            }
+            return 0;
+        }
+
         public async Task<IActionResult> Index()
         {
+            int userId = GetUserId();
+            ViewBag.UsuarioID = userId;
+            if (userId == 0)
+            {
+                return Forbid();
+            }
+
             var caixaItems = await _context.Caixa.ToListAsync();
 
             var entradas = _context.Caixa
-               .Where(c => c.Categoria == CategoriaMovimentacao.Entrada)
+               .Where(c => c.Categoria == CategoriaMovimentacao.Entrada && c.UsuarioID == userId)
                .Sum(c => c.Valor);
 
             var saidas = _context.Caixa
-                .Where(c => c.Categoria == CategoriaMovimentacao.Saída)
+                .Where(c => c.Categoria == CategoriaMovimentacao.Saída && c.UsuarioID == userId)
                 .Sum(c => c.Valor);
 
             ViewBag.Entradas = entradas;
@@ -38,21 +55,49 @@ namespace AGENDAHUB.Controllers
 
         public IActionResult SearchByDate(DateTime dataInicio, DateTime dataFim)
         {
-            // FALTA CRIAR
+            int userId = GetUserId();
+            ViewBag.UsuarioID = userId;
+            if (userId == 0)
+            {
+                return Forbid();
+            }
+
+            //Falta Criar
+
             return View();
         }
 
         // Função para exibir a tela de cadastro de movimentação de caixa
         public IActionResult Create()
         {
+            int userId = GetUserId();
+            ViewBag.UsuarioID = userId;
+
+            if (userId == 0)
+            {
+                return Forbid();
+            }
+
             return View();
         }
 
-        // Resposta HTTP para adicionar uma movimentação de caixa no banco de dados
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(Caixa caixa)
         {
+            var userId = GetUserId();
+            ViewBag.UsuarioID = userId;
+
+            if (userId == 0 || userId == null)
+            {
+                return Forbid();
+            }
+
+            if (!_context.Usuarios.Any(u => u.Id == caixa.UsuarioID))
+            {
+                return NotFound("Usuário não encontrado.");
+            }
+
             if (ModelState.IsValid)
             {
                 _context.Caixa.Add(caixa);
@@ -62,19 +107,30 @@ namespace AGENDAHUB.Controllers
             return View(caixa);
         }
 
+
         // Função para exibir a tela de edição de uma movimentação de caixa
         public async Task<IActionResult> Edit(int? id)
         {
+            int userId = GetUserId();
+
+            if (userId == 0)
+            {
+                return Forbid();
+            }
+
             if (id == null)
             {
                 return NotFound();
             }
 
-            var caixa = await _context.Caixa.FindAsync(id);
+            var caixa = await _context.Caixa
+                .FirstOrDefaultAsync(m => m.ID_Caixa == id && m.UsuarioID == userId);
+
             if (caixa == null)
             {
                 return NotFound();
             }
+
             return View(caixa);
         }
 
@@ -83,9 +139,11 @@ namespace AGENDAHUB.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, Caixa caixa)
         {
-            if (id != caixa.ID_Caixa)
+            int userId = GetUserId();
+            ViewBag.UsuarioID = userId;
+            if (userId == 0 || id != caixa.ID_Caixa || caixa.UsuarioID != userId)
             {
-                return NotFound();
+                return Forbid();
             }
 
             if (ModelState.IsValid)
@@ -108,40 +166,59 @@ namespace AGENDAHUB.Controllers
                 }
                 return RedirectToAction("Index");
             }
+
             return View(caixa);
         }
 
         // Função para exibir a tela de detalhes de uma movimentação de caixa
         public async Task<IActionResult> Details(int? id)
         {
+            int userId = GetUserId();
+            ViewBag.UsuarioID = userId;
+            if (userId == 0)
+            {
+                return Forbid();
+            }
+
             if (id == null)
             {
                 return NotFound();
             }
 
             var caixa = await _context.Caixa
-                .FirstOrDefaultAsync(m => m.ID_Caixa == id);
+                .FirstOrDefaultAsync(m => m.ID_Caixa == id && m.UsuarioID == userId);
+
             if (caixa == null)
             {
                 return NotFound();
             }
+
             return View(caixa);
         }
 
         // Função para exibir a tela de confirmação de exclusão de uma movimentação de caixa
         public async Task<IActionResult> Delete(int? id)
         {
+            int userId = GetUserId();
+            ViewBag.UsuarioID = userId;
+            if (userId == 0)
+            {
+                return Forbid();
+            }
+
             if (id == null)
             {
                 return NotFound();
             }
 
             var caixa = await _context.Caixa
-                .FirstOrDefaultAsync(m => m.ID_Caixa == id);
+                .FirstOrDefaultAsync(m => m.ID_Caixa == id && m.UsuarioID == userId);
+
             if (caixa == null)
             {
                 return NotFound();
             }
+
             return View(caixa);
         }
 
@@ -150,7 +227,21 @@ namespace AGENDAHUB.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var caixa = await _context.Caixa.FindAsync(id);
+            int userId = GetUserId();
+            ViewBag.UsuarioID = userId;
+            if (userId == 0)
+            {
+                return Forbid();
+            }
+
+            var caixa = await _context.Caixa
+                .FirstOrDefaultAsync(m => m.ID_Caixa == id && m.UsuarioID == userId);
+
+            if (caixa == null)
+            {
+                return NotFound();
+            }
+
             _context.Caixa.Remove(caixa);
             await _context.SaveChangesAsync();
             return RedirectToAction("Index");
